Moved signup into onboarding_api and added formatting.
