+++ bctw-api/src/auth/authHelpers.ts
@@ -1,59 +1,59 @@
-import { Audience } from "../types/userRequest";
+import { Audience } from '../types/userRequest';
 
 export const ROUTE_AUDIENCES: { [key: string]: Audience[] } = {
-    'GET:/get-template': [],
-    'GET:/get-critters': ['BCTW'],
-    'GET:/get-critter-tracks': ['BCTW'],
-    'GET:/get-pings-estimate': ['BCTW'],
-    'GET:/get-animals': ['BCTW'],
-    'GET:/get-attached-historic': ['BCTW'],
-    'GET:/get-animal-history/:animal_id': ['BCTW'],
-    'POST:/upsert-animal': ['BCTW'],
-    'GET:/get-all-collars': ['BCTW'],
-    'GET:/get-collars-and-deviceids': ['BCTW'],
-    'GET:/get-assigned-collars': ['BCTW'],
-    'GET:/get-available-collars': ['BCTW'],
-    'GET:/get-assignment-history/:animal_id': ['BCTW'],
-    'GET:/get-collar-history/:collar_id': ['BCTW'],
-    'GET:/get-collar-vendors': ['BCTW', 'SIMS'],
-    'POST:/upsert-collar': ['BCTW'],
-    'POST:/attach-device': ['BCTW'],
-    'POST:/unattach-device': ['BCTW'],
-    'POST:/update-data-life': ['BCTW'],
-    'GET:/permission-request': ['BCTW'],
-    'GET:/permission-history': ['BCTW'],
-    'POST:/submit-permission-request': ['BCTW'],
-    'POST:/execute-permission-request': ['BCTW'],
-    'GET:/get-user': ['BCTW'],
-    'GET:/get-users': ['BCTW'],
-    'GET:/get-user-role': ['BCTW'],
-    'POST:/signup': ['SIMS_SERVICE'], // Only the SIMS service can signup users directly.
-    'POST:/add-user': ['BCTW'],
-    'GET:/get-onboard-status': [], // Open to all common-realm users.
-    'GET:/onboarding-requests': ['BCTW'],
-    'POST:/submit-onboarding-request': [],
-    'POST:/handle-onboarding-request': ['BCTW'],
-    'GET:/get-critter-access/:user': ['BCTW'],
-    'POST:/assign-critter-to-user': ['BCTW'],
-    'POST:/add-udf': ['BCTW'],
-    'GET:/get-udf': ['BCTW'],
-    'GET:/get-user-alerts': ['BCTW'],
-    'GET:/test-alert-notif': ['BCTW'],
-    'POST:/update-user-alert': ['BCTW'],
-    'GET:/get-code': ['BCTW'],
-    'GET:/get-code-headers': ['BCTW'],
-    'GET:/get-code-long-desc': ['BCTW'],
-    'POST:/export': ['BCTW'],
-    'POST:/export-all': ['BCTW'],
-    'POST:/import-xlsx': ['BCTW'],
-    'POST:/import-finalize': ['BCTW'],
-    'POST:/deploy-device': ['SIMS'], // Only SIMS users may do this.
-    'POST:/import-xml': ['BCTW'],
-    'POST:/import-telemetry': ['BCTW'],
-    'GET:/get-collars-keyx': ['BCTW'],
-    'POST:/fetch-telemetry': ['BCTW'],
-    'DELETE:/:type': ['BCTW'],
-    'DELETE:/:type/:id': ['BCTW'],
-    'GET:/:type/:id': ['BCTW'],
-    'GET:/health': ['BCTW', 'SIMS'],
-}
+  'GET:/get-template': [],
+  'GET:/get-critters': ['BCTW'],
+  'GET:/get-critter-tracks': ['BCTW'],
+  'GET:/get-pings-estimate': ['BCTW'],
+  'GET:/get-animals': ['BCTW'],
+  'GET:/get-attached-historic': ['BCTW'],
+  'GET:/get-animal-history/:animal_id': ['BCTW'],
+  'POST:/upsert-animal': ['BCTW'],
+  'GET:/get-all-collars': ['BCTW'],
+  'GET:/get-collars-and-deviceids': ['BCTW'],
+  'GET:/get-assigned-collars': ['BCTW'],
+  'GET:/get-available-collars': ['BCTW'],
+  'GET:/get-assignment-history/:animal_id': ['BCTW'],
+  'GET:/get-collar-history/:collar_id': ['BCTW'],
+  'GET:/get-collar-vendors': ['BCTW', 'SIMS'],
+  'POST:/upsert-collar': ['BCTW'],
+  'POST:/attach-device': ['BCTW'],
+  'POST:/unattach-device': ['BCTW'],
+  'POST:/update-data-life': ['BCTW'],
+  'GET:/permission-request': ['BCTW'],
+  'GET:/permission-history': ['BCTW'],
+  'POST:/submit-permission-request': ['BCTW'],
+  'POST:/execute-permission-request': ['BCTW'],
+  'GET:/get-user': ['BCTW'],
+  'GET:/get-users': ['BCTW'],
+  'GET:/get-user-role': ['BCTW'],
+  'POST:/signup': ['SIMS_SERVICE'], // Only the SIMS service can signup users directly.
+  'POST:/add-user': ['BCTW'],
+  'GET:/get-onboard-status': [], // Open to all common-realm users.
+  'GET:/onboarding-requests': ['BCTW'],
+  'POST:/submit-onboarding-request': [],
+  'POST:/handle-onboarding-request': ['BCTW'],
+  'GET:/get-critter-access/:user': ['BCTW'],
+  'POST:/assign-critter-to-user': ['BCTW'],
+  'POST:/add-udf': ['BCTW'],
+  'GET:/get-udf': ['BCTW'],
+  'GET:/get-user-alerts': ['BCTW'],
+  'GET:/test-alert-notif': ['BCTW'],
+  'POST:/update-user-alert': ['BCTW'],
+  'GET:/get-code': ['BCTW'],
+  'GET:/get-code-headers': ['BCTW'],
+  'GET:/get-code-long-desc': ['BCTW'],
+  'POST:/export': ['BCTW'],
+  'POST:/export-all': ['BCTW'],
+  'POST:/import-xlsx': ['BCTW'],
+  'POST:/import-finalize': ['BCTW'],
+  'POST:/deploy-device': ['SIMS'], // Only SIMS users may do this.
+  'POST:/import-xml': ['BCTW'],
+  'POST:/import-telemetry': ['BCTW'],
+  'GET:/get-collars-keyx': ['BCTW'],
+  'POST:/fetch-telemetry': ['BCTW'],
+  'DELETE:/:type': ['BCTW'],
+  'DELETE:/:type/:id': ['BCTW'],
+  'GET:/:type/:id': ['BCTW'],
+  'GET:/health': ['BCTW', 'SIMS'],
+};
