+++ bctw-api/src/auth/authorization.ts
@@ -6,10 +6,11 @@ import {
   query,
 } from '../database/query';
 import { UserRequest } from '../types/userRequest';
-import { IUserInput, eUserRole } from '../types/user';
 import { ROUTE_AUDIENCES } from './authHelpers';
 
-const getRegistrationStatus = async (keycloakId: string): Promise<boolean> => {
+export const getRegistrationStatus = async (
+  keycloakId: string
+): Promise<boolean> => {
   const sql = constructFunctionQuery(fn_get_user_id, [keycloakId]);
   const { result } = await query(sql);
   const isRegistered =
@@ -42,49 +43,5 @@ export const authorizeRequest = async (
     }
   }
 
-  next();
-};
-
-const signupUser = async (req: UserRequest): Promise<void> => {
-  const { keycloakId, domain, email, givenName, familyName } = req.user;
-  const user = {
-    [domain]: keycloakId,
-    email,
-    firstname: givenName,
-    lastname: familyName,
-  } as Record<keyof IUserInput, unknown>;
-  const sql = constructFunctionQuery('upsert-user', [
-    'system',
-    user,
-    eUserRole.user,
-  ]);
-  const { result, error, isError } = await query(sql, '', true);
-  if (isError) {
-    throw new Error(error.message);
-  }
-};
-
-export const signup = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-): Promise<void> => {
-  try {
-    await signupUser(req as UserRequest);
-  } catch (error) {
-    console.log(error);
-    res.status(500).send((error as Error).message);
-    return;
-  }
-
-  // Confirm successful registration
-  const registered = await getRegistrationStatus(
-    (req as UserRequest).user.keycloakId
-  );
-  if (!registered) {
-    res.status(500).send('Failed to register user');
-    return;
-  }
-  (req as UserRequest).user.registered = registered;
   next();
 };
